+++ src/lib/server/db/actions/importActions.ts
@@ -18,7 +18,7 @@ import {
 import { updatedTime } from './helpers/misc/updatedTime';
 import { eq, and, getTableColumns, desc, sql } from 'drizzle-orm';
 import { tActions } from './tActions';
-import { filterNullUndefinedAndDuplicates } from '../../../../routes/(loggedIn)/journals/filterNullUndefinedAndDuplicates';
+import { filterNullUndefinedAndDuplicates } from '$lib/helpers/filterNullUndefinedAndDuplicates';
 import type { ZodSchema } from 'zod';
 import { importTypeEnum, type importTypeType } from '$lib/schema/importSchema';
 import {
